Implement program main loop
The bash-clone interpreter can now be ran from command line
in two ways.
1) basic interactive mode
./bash-clone <init-script-file>

It is necessary to provide an initialization script.

2) script mode
./bash-clone <init-script-file> -c <script-file>

The difference between these modes is that in interactive mode
the program waits for user input and prints encourage symbol, while
in non-interactive mode it finishes execution after reaching EOF.